Split SpriteRenderable into its own file.
